fix: Reimplement timestamp computation
This is a partial improvement to address inconsistencies in browsers and
how they implement performance.now().

It mitigates observing timestamps in the past for pageload transactions,
but gives no guarantees to navigation transactions.

Depending on the platform, the clock used in performance.now() may stop
when the computer goes to sleep, creating an ever increasing skew. This
skew is more likely to manifest in navigation transactions.